js/src/xpconnect/tests/mochitest/test_bug448587.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=448587
-->
<head>
  <title>Test for Bug 448587</title>
  <script type="text/javascript" src="/MochiKit/packed.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=448587">Mozilla Bug 448587</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript; version=1.8">

/** Test for Bug 448587 **/
let (f = function() {}) {
    ok(new XPCSafeJSObjectWrapper(f).prototype == f.prototype ,
       "XPCSafeJSObjectWrapper allows getting the 'prototype' property");
}

</script>
</pre>
</body>
</html>
